ceph: use uuid when creating zombie storage volumes
Ceph is similar to zfs in that storage volumes need to be kept around in case
they have dependent snapshots. That is:

- <storage-volume-1>
- snapshot(<storage-volume-1) = <snap1>
- protect(<snap1>)            = <snap1_protected>
- delete(<storage-volume-1>)

will fail since <snap1_protected> is dependent. Because of this the way I
implemented the ceph driver it will map the delete() operation onto a rename()
operation:

- delete(<storage-volume-1>) -> rename(<storage-volume-1>) = zombie_<storage-volume-1>

But consider the following:

lxc init images:alpine/edge a = <storage-volume-a>                                  /* succeeds */
lxc copy a b                  = <storage-volume-b> depends <storage-volume-a>       /* succeeds */
lxc delete a                  = <storage-volume-a> rename zombie_<storage-volume-a> /* succeeds */
lxc init images:alpine/edge a = <storage-volume-a>                                  /* succeeds */
lxc copy a c                  = <storage-volume-c> depends <storage-volume-a>       /* succeeds */
lxc delete a                  = <storage-volume-a> rename zombie_<storage-volume-a> /* fails    */

To avoid this we suffix each storage volume with a UUID so the sequence above becomes:

lxc init images:alpine/edge a = <storage-volume-a>                                         /* succeeds */
lxc copy a b                  = <storage-volume-b> depends <storage-volume-a>              /* succeeds */
lxc delete a                  = <storage-volume-a> rename zombie_<storage-volume-a>_<uuid> /* succeeds */
lxc init images:alpine/edge a = <storage-volume-a>                                         /* succeeds */
lxc copy a c                  = <storage-volume-c> depends <storage-volume-a>              /* succeeds */
lxc delete a                  = <storage-volume-a> rename zombie_<storage-volume-a>-<uuid> /* succeeds */

Closes #3780.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>