lvm: disallow thinpools + btrfs filesystems
When a thinpool is used for the lvm storage pool we use logical volume
snapshots. This will cause filesystems to have identical UUIDs which is a
problem. xfs allows changing the UUID of a filesystem easily. The excellent
toolsuite that btrfs provides however, will not let you change the UUID of a
btrfs device **non-interactively**. We could start to carry our own btrfs UUID
implementation but this is a major pain as it means writing at least ~500 lines
of C code. Since we haven't seen anyone complain or major users of lvm + btrfs
I say we simply disallow thinpool-based lvm storage pools with btrfs
filesystems.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>