network: add netns_getifaddrs() infrastructure
This commit introduces my concept of a network namespace aware
getifaddrs(), i.e. netns_getifaddrs().
This presupposes a kernel that supports IF{L}A_TARGET_NETNSID.

I am also introducing:

struct netns_ifaddrs {
	struct netns_ifaddrs *ifa_next;

	// Can - but shouldn't be - NULL.
	char *ifa_name;

	// This field is not present struct ifaddrs
	int ifa_ifindex;
	unsigned ifa_flags;

	// This field is not present struct ifaddrs
	int ifa_mtu;
	struct sockaddr *ifa_addr;
	struct sockaddr *ifa_netmask;
	union {
		struct sockaddr *ifu_broadaddr;
		struct sockaddr *ifu_dstaddr;
	} ifa_ifu;

	// If you don't know what this is for don't touch it.
	void *ifa_data;
};

which is a superset of struct ifaddrs. It contains additional
information such as the mtu and ifindex of the corresponding network
devices. Note though that the field ordering is not is different too. So
don't get any ideas of using memcpy() to copy from an old struct ifaddrs
into a struct netns_ifaddrs.
I would suggest using to either use a custom struct + map:

type NetnsGetifaddrs struct {
	// Sensible fields for general stuff we need
}

var ifaddrs map[int]NetnsGetifaddrs

where the map keys are the ifindices of the interfaces or to translate
it into the standard GO struct for this type of information.

This commits also adds dumping infrastructure. It is not needed for LXD
and once the implemenation is complete it can be removed. But it is very
useful to debug the returned C struct itself and might be worth keeping.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>